backend/requirements.txt
# Django Core
Django>=5.2.8
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
mysqlclient>=2.2.0

# Environment
python-dotenv>=1.0.0

# AI/MLtre
openai>=2.8.1
openai-agents>=0.6.1
langchain-core>=1.1.0
langchain-openai>=1.1.0
langchain-text-splitters>=1.0.0
tavily-python>=0.7.13
langfuse>=2.0.0

# Data Processing (already installed)
pandas>=2.3.3
numpy>=2.3.5
openpyxl>=3.1.5

# Vector Database (already installed)
qdrant-client>=1.16.1

# Utils (already installed)
PyMySQL>=1.1.2
requests>=2.32.5
python-multipart>=0.0.20
